Rename progress to state in GrandExchangeOffer
The field formerly known as "progress" encoded what state the offer was
in which is a weird 4-bit thing that includes information about whether
the offer is a buy or sell offer, whether it's finished, and some other,
less well-defined info about (what seems to be) the offer's live-ness.
How close an offer is to being finished (ie its progress) is only really
based on how many items have bought/sold out of the total amount
requested/offered.