Add movable overlay support
This overhauls most of the overlay renderer to do so. There are 4 (or 5) "snap corners" that let overlays change their position, or you can set a custom location which is at a set x/y.

Calculate which layer overlays are on up front and build separate lists for each layer. The fixed-point under-widget overlays are moved to above-widget so that they work when moved above the interfaces.