banktags: Don't use cs2 to add menu options
The mobile bank interface uses the same index as we do for menu option, and fully saturates a widget's 10 available menu options (Excluding type 2 widget options, which are special). This means we cannot use the cs2 opcode to add a menu option without overwriting an existing one, so we have to add it at menu build time.