config service: convert backing database to use mongodb
This is a much more natural fit for the config service which is really a
key value store, and has been abusing SQL. This will let us expand the
config stuff later to support profiles and per-account config values.

I have left in the SQL code for now so config changes are still being
sent there in the event of catastrophic mongodb failure.