Remove PVP references to HIGH_RISK world type
As described in 695ca9c7c, HIGH_RISK is not a PVP world type.

Note: this will not affect the world hopper plugin, as it both
explicitly references the high risk world type, and references the
http-api WorldType implementation, not the runelite-api WorldType which
is being modified in this commit.

Fixes runelite/runelite#8605
Fixes runelite/runelite#8624