attack styles: fix hide styles on weapon change
Hiding attack styles primarily relies on a subscription to
onWidgetHiddenChanged but when swapping between bludgeon and bow this event
fires before onVarbitChanged which means that when swapping quickly between
the two, the longrange option will not be hidden. Since the plugin uses info about
the weapon to determine in which skill xp will be gained and ultimately, which widgets
should be hidden, we need an extra call to ProcessWidgets after a weapon switch to ensure
expected behavior.