runelite-client: Make RuneScape Bold actually Bold
This changes the font name and family for runescape.ttf to just
`RuneScape`, and the family of runescape_bold.ttf to `RuneScape`, with
the name `RuneScape Bold`. It also sets OS/2 weight and style to Bold.

This allows swing's HTML viewer to switch between these fonts naturally
with just a font-weight style, or a <strong> element