kourendlibrary: Indicate books not in inventory
This commit changes the Kourend library panel to show a different color
for books which are not the target based on whether they are in the
player's inventory or not. Previously, the target book would be shown in
green and all other books would be listed in white. Now, non-target
books in the player's inventory are shown in white and those the player
does not have are shown in orange.

Co-authored-by: Jordan Atwood <jordan.atwood423@gmail.com>