client: block extra mouse keys
The key remapping plugin has no way to operate being on and also not
blocking chat input, due to the other features of the plugin all doing
key remaps and thus are required to know whether or not you are trying
to type, or using a remapped key.

This moves the blocking to the core, which we think won't affect many
users anyway, and those that it does can just remap their mouse keys.