interfacestyles: update widget bounds before every render
this reverts commit 9b365e2c542afa676f9f82445222473380408467

We need to run this every frame because the client can update the widget
bounds cache without calling [proc,toplevel_resize]. This is easily
reproducible by resizing the client in fixed mode, which causes the
the tli to be revalidated, but because its in fixed mode it doesn't run
the resize listener because the bounds don't actually change.

We use BeforeRender instead of ClientTick because ClientTick is posted
before cs2s are evaluated each tick, so if [proc,toplevel_resize] (or
any other cs2 that can touch these widgets) is called the widget would
be in the vanilla position for the next frame.