Add all other pearl rod animations
There are 2 animations for each pearl rod (except Oily). Usually it
seems that only one is used when actually fishing, which were the ones
present before this commit. However, it seems that Jagex is not
consistent about what animations are used, as Anglerfish fishing uses
the other animation of the normal pearl rod. Given that the other two
exist, and they're technically usable (and can be used as the starting
animation), it's only a matter of time before Jagex accidentally gets
them the wrong way round again, so this is a pre-emptive measure. It
also means that the UI won't show "Not Fishing" for a split second when
casting out.