chat history: fix replaying runelite-formatted messages
This stores the underlying message node, which may be updated much later with a
new message value if a command is used, and replays that. Previously
this was storing and replaying the raw RL formatted message, which is
why colors would be stripped off of RL-formatted messages on replay.

This also no longer triggers chat message events from history replay,
which fixes commands incorrectly refiring from replay (and thus the nbsp
hack), and fixes the chat notifier sending notifications for replayed
messages.