timers: Add cannon timer for each parts' setup (#13856)
Due to certain new PVM methods, there are situations in which it is
useful to set up a partial cannon to indicate a world which is taken, or
to be placed in a strategic location for a red-click action. For this
reason, it is useful to display the cannon decay timer for all stages of
cannon setup.