Fix race with client startup and low detail/chat icon loading
Since 95b8daaa81694a62998097758eca24a7f6a94cf6 moved client startup to
be much earlier, the plugins can now start after the client is already
at the login screen. This causes the low detail plugin and chat icon
manager to not initialize correctly as they are waiting on the event
for the login screen.