tooltip: don't set overlay bounds when rendering
The overlay was storing the previous frames tooltip width and height in
the overlay bounds, to use for the next frame to position the overlay
correctly so that it did not get pushed off canvas.

This requires the tooltip overlay to have a non-empty bounds, which is
not something dynamic/tooltip overlays usually support. The overlay
renderer translates the g2d to the given bounds prior to rendering.

The reason this worked was it always kept x/y=0, causing the translation
to not happen.

However this incorrectly caused the overlay to be picked when testing
overlay bounds against the mouse position when up near the corner of the
screen. This would then cause the player to pick the tooltip overlay,
despite the overlay not being rendered there.