loottracker: sync config on configsync
Previously there were two conditions where loottracker would flush kills
to the wrong config:

1) SessionClose is called after the session has been fully destroyed and
   the config swapped out with the offline config

2) RuneScapeProfileChanged can be called due to a config swap from
   switchSession, which is after the config client has the new session
   and the new config loaded

RuneScapeProfileChanged has multiple other callsites due to profile
changes not due to a config swap, so the logic of swapping the active
profile must be kept for that.

ConfigSync is called correctly during the above two scenarios and is
also on a core 5 minute timer, so it can replace the current scheduled
task.