hiscore panel: repaint entire panel on lookup
Updating the individual labels is treated as separate draws by d3d, and
the d3d thread performs a swapBuffers() for each. With drivers
configured to apply FPS limits, each swap is treated as a frame, making
it very easy to exceed the frame limit. This causes the EDT to spend
most of its time in D3DScreenUpdateManager::runUpdateNow waiting for the
buffer swap to complete, causing visible lag.

Marking the entire panel for repaint fixes it by reducing the number of
draws.