loot tracker: fix tracking whisperer deaths v3
This almost worked before, but since player update happens prior to npc update,
if the player died and was teleport to spawn, the snapshot region was set to
the spawn region. So use the npc location, but it must be converted from
instance coordinates, since the fight takes place at the intersection of 4 map
corners.

Also fix bug where fromLocalInstance was calling the wrong overload. Client
instance chunks/basexy should always be preferred over Scene, since the Actors
coordiantes are always relative to it and not to the Scene.