loot tracker: Fix The Leviathan loot tracking
The Leviathan drops its loot on the tile the player was standing on when
it begins its death animation. However, because the arena is a
single-player instance, there is no reason not to scan the entire floor.
This will cover any future changes to loot drop position and possible
discrepancies in location for additional dropped items like brimstone
keys.