gpu: remove sin/cos table from UBO
This was added in aef933eb0cbe36b441bbc019a8ef42e4952d5b13 due to
inconsistencies in the computed sin/cos values causing graphical issues
on some cards. However the sin/cos table is no longer used for the
projection, and instead only for computing face distance for face
sorting, so I think it is safe to remove.