bank tags: fix scrolling the tag tab
Having a separate layer for the tag tab tab would require either a
separate scrollbar or adjusting the bank scrollbar dynamically between
the two layers.

This is somewhat tricky, so instead just use the previous behavior of
creating the tag tab tab items after the other bank components (this is
so they dont interfere with the normal bank items which have their
clickmask set to send ops).

Move the rebuildTagTabTab() call that was running post bankmain_build to
be instead pre bankmain_finish so that it runs prior to the scrollbar
calculation.