slayer: Fix initial amount when task changes
This is most commonly encountered when players are assigned a Tzhaar
task, but elect to upgrade it to a Jad or Zuk task. When this happens,
the SLAYER_TASK_SIZE and SLAYER_TASK_CREATURE varps are set when
starting the assignment dialog, and are later changed after opting in to
the task change. The plugin had code which would preserve the previous
initial amount (as of d088d39bd769a5c337588838d45d29cb54fa771c) but
should no longer be needed as we can update it directly from var changes
as opposed to tracking it ourselves.