ui: readd window opacity setting
The undecorated frame requirement appears to not be real, at least on Windows.

AWT applies translucency via setting WS_EX_LAYERED to make the window layered and then changes the opacity of the layered window with SetLayeredWindowAttributes.

This appears to work regardless of the frame state, even with custom chrome off.