loot manager: fix nightmare tracking
The loot area calculation was moved in
95f34bc90daa31faed403802a9fcb660088ebbdc, and broke because it needs to
be called at the same tick the loot spawns as it uses the player
location.

Move back the calculation of the loot area to when the loot spawns,
unless already set, by eg. the duke tracking.

Also change the delayed loot timings to be tick-perfect, fix the wall
beast timing, and use getDropLocations for them()

Co-authored-by: umer-rs <72721490+umer-rs@users.noreply.github.com>