bank tags: add layouts
This incorporates bank tag layouts as seen by the popular bank tag
layouts plugin. Thanks to @geheur for their assistance with development
and testing, and for the original idea.

There are also several new APIs for plugins to manage tags, tabs, and
layouts.

Example API usage to open a custom layout:

  Layout l = new Layout();
  l.setItemAtPos(ItemID.ROBIN_HOOD_HAT, 4);
  BankTag bt = new BankTag()
  {
      public boolean contains(int itemId)
      {
          return l.count(itemId) > 0;
      }

      public Layout layout()
      {
          return l;
      }
  };
  bankTagsService.openBankTag(bt);