Prayer Helper
using the AttackTimer plugin off the Plugin Hub, I have added the ability for prayers to enable using the "Lazy Flick" method, and also enable normally if you dont want it to flick. This is all automatic based on attack style and weapon speed etc.

Option to draw the overlay for the tick